fix vfolder deletion codes
