+++ src/ai/backend/manager/clients/storage_proxy/BUILD
@@ -1,0 +1,1 @@
+python_sources(name="src")
