+++ src/ai/backend/manager/data/user/types.py
@@ -1,13 +1,41 @@
+from __future__ import annotations
+
+import enum
 from dataclasses import dataclass, field
 from datetime import datetime
-from typing import Any, Optional, Self
+from typing import Any, Optional, Self, override
 from uuid import UUID
 
 from sqlalchemy.engine import Row
 
 from ai.backend.common.types import AccessKey
-from ai.backend.manager.models.user import UserStatus
 from ai.backend.manager.types import Creator
+
+
+class UserStatus(enum.StrEnum):
+    """
+    User account status.
+    """
+
+    ACTIVE = "active"
+    INACTIVE = "inactive"
+    DELETED = "deleted"
+    BEFORE_VERIFICATION = "before-verification"
+
+    @override
+    @classmethod
+    def _missing_(cls, value: Any) -> Optional[UserStatus]:
+        assert isinstance(value, str)
+        match value.upper():
+            case "ACTIVE":
+                return cls.ACTIVE
+            case "INACTIVE":
+                return cls.INACTIVE
+            case "DELETED":
+                return cls.DELETED
+            case "BEFORE-VERIFICATION" | "BEFORE_VERIFICATION":
+                return cls.BEFORE_VERIFICATION
+        return None
 
 
 @dataclass
